Backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Search
faiss-cpu>=1.8.0

# AI
groq>=0.4.0

# Embeddings
sentence-transformers==2.3.1

# Utilities
python-dotenv==1.0.1
pydantic>=2.6.1,<3
pydantic-settings>=2.1.0

# Async & HTTP
httpx==0.26.0
anyio==4.2.0
starlette==0.36.3

# MCP Server
mcp>=1.0.0
sse-starlette==1.8.2
